requirements.txt
discord.py
langchain
langchain_community
Pillow
pydantic
pytesseract
fastapi 
uvicorn 
faiss-cpu         
pydantic
transformers
langchain_text_splitters
torch
pymongo
bs4
gradio
lxml
argparse
streamlit
langchain-groq
langchain_core
PyPDF2
python-dotenv
